requirements.txt

certifi==2019.3.9
chardet==3.0.4
clarifai==2.6.1
Django==2.1.7
django-markdown-deux==1.0.5
django-pagedown==1.0.6
docutils==0.14
gunicorn==19.9.0
idna==2.8
Jinja2==2.10
Markdown==3.0.1
markdown2==2.3.7
MarkupSafe==1.1.1
MeaningCloud-python==1.1.1
mistune==0.8.4
mock==2.0.0
mpmath==1.1.0
nbconvert==5.4.1
nbformat==4.4.0
nose==1.3.7
parso==0.3.4
pbr==5.1.3
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
statistics==1.0.3.5
sympy==1.4
whitenoise==4.1.2
widgetsnbextension==3.4.2
